[Refactor]: AcceptanceTest 디렉토리 변경
